--- src/components/footer/footer.jsx
@@ -1,7 +1,0 @@
-export default function Footer() {
-    return (
-        <footer className="bg-blue-500 text-center text-white p-4">
-            <p>alBz© 2024. MIT license.</p>
-        </footer>
-    )
-}
