+++ src/components/layout/footer/index.jsx
@@ -1,0 +1,7 @@
+export default function Index() {
+    return (
+        <footer className="bg-black text-center text-white p-4">
+            <p>alBz© 2024. MIT license.</p>
+        </footer>
+    )
+}
