+++ src/components/layout/header/index.jsx
@@ -2,13 +2,13 @@ import {
     NavigationMenu,
     NavigationMenuItem,
     NavigationMenuList,
-} from "@/components/ui/navigation-menu";
+} from "@/components/ui/navigation-menu.jsx";
 
 const publicPath = import.meta.env.VITE_BASE_URL || '/';
 
-export default function Header() {
+export default function Index() {
     return (
-        <header className="bg-blue-500 p-3 min-h-[10px]">
+        <header className="bg-black p-3 min-h-[10px]">
             <NavigationMenu className="min-w-full">
                 <NavigationMenuList className="flex space-x-4">
                     <NavigationMenuItem>
