+++ src/routes/home.jsx
@@ -1,13 +1,13 @@
-import Layout from "../layout.jsx";
+import Index from "../components/layout/index.jsx";
 import { Button } from "@/components/ui/button"
 
 export default function Home() {
     return (
         <>
-            <Layout>
+            <Index>
                 <h1 className="mt-2">Home</h1>
                 <Button>Click me</Button>
-            </Layout>
+            </Index>
         </>
     )
 }
